src/requirements.txt
sounddevice
pillow
requests
shazamio
scipy
matplotlib
scikit-learn
